feat(repeatWhen): add higher-order lettable version of repeatWhen
